Create Identity Management Stubs
To provision a Kubernetes cluster (with CAPO) we simply need a cloud
config and the external network ID (but maybe not any more...).  We also
need to make note of the project and user IDs so when the cluster is
deprovisioned we can clean them up.  We wrap all this as an identity
object, the cluster creation requests an identity, gets back all the
credentials it needs to create the cluster.  The k8s controller can
remember the identity ID as an annotation, and when the time comes
delete the identity.  Pertinent information is also saved as a custom
resource.  We can even extend the API to create a provider network in
the context of the identity, or a compute instance... you can see where
this is going.  All in a nice provider agnostic way!