Make Everything Project Scoped (#15)
We will need to start locking down APIs to specific projects that
require role permissions to be present, rather than at the organization
level, so do this now rather that risk later pain.  It does have the
benefit that we can now scope regions and other jazz to a specific
project in an organizations etc.