Finish Identity View
This has identified a number of UX problems, in that we can list
identities for an organization, but then we need to get regions per
identity (project) to correctly scope regions to translate from ID to
name.  I suspect that perhaps regions should be defined at the
organization level - not the project, identities should have a regionID
property much like clusters, then we can handle those in a geenric
fashion and promote region scoped metadata to a top level core thing.