Add Server Group Cleanup (#45)
Nova just orphans things on project deletion, which leaves a bunch of
server groups hanging around in the aether, so we should play nice and
clean these up for it, and make life easier on our platform teams.  Also
while here standarize the trace output on service name (as given in
openstack versions show) and add a method to make debug easier.