Identity Controller Boilerplate
At present, when provisioning in a cluster, it blocks a long time
waiting for OpenStack to do what it needs to do.  This has a knock on
effect for UX in that said delay is visible in the UI and CLI tools.
This lays down the groundwork to make this asychronous so we don't hold
up uer interaction, but we do cluster provisioning.  This is a template
also for phyiscal network provisioning that takes even longer, and
causes request timeouts that need to be avoided.