Add Quota Management Support (Part 1)
Without this, you'll not get very far with Openstack's defaults!  This
adds in an endpoint that can read and update quotas.  These are
explciitly scoped to the identity (as that maps to an OpenStack project
where quotas reside), and implicitly to the client e.g.
unikorn-kubernetes/unikorn-baremetal, so they can independently affect
quotas in a shared identity without having to know about one another.

This first part adds in the necessary APIs that allow quotas to be
created by a client, trigger an identity reconcile and are implicitly
deleted on identity deletion.