Automate Image Selection
Prior to this change, we let the user select what images were presented
to the user.  Since the time when this was written, we have released a
versioned specification that controls what consitutes a usable image,
taking into account UX requirements, and those of various higher level
services e.g. Kubernetes.  Remove the capability for a user to manaually
select images and instead use the schema defined by the specification.
The signing key is still available to verify the provenance of an image
on platforms where users can, legitmately, create a valid image
compatible with the schema.

Fixes #112