Refactor Handler (#145)
Handlers should not be used in line, they should delegate to a specific
client so the logic could be shared, plus everything works in a
consistent fashion then which makes cross repository changes e.g.
filtering etc. a lot less painful.

Also, while we are in the area, I spotted a serious security bug where
when accessing a resource directly by ID (e.g. GET/PUT/DELETE), we
weren't checking that resource organization and project matched those
that we specified at the API (and thus sanitized by RBAC).  The key part
here is if you know an ID, and have access to any project, you can
delete any resource.  This is a problem specific to the region
controller as everything else uses namespaces for segregation.  It's
also not really much of a problem as nothing directly accesses the
region controller.

Fixes #144