Simplify Server API
The API was tailored in a way to possibly provide an image selector, but
that's not necessary, and in most cases undesirable and it may lead to
accidental upgrades/deletions, so KISS.  This is a breaking change, but
it's a private API anyway so will be handled in compute.

Fixes #141.