Fix Security Group Deletion
As it stands security groups will be deleted when identities are, and
will error until all server ports stop using them.  This adds server
references to security groups (and keeps then in sync) to mostly prevent
this (there are abviously some races, but less so now).  This will allow
security groups to be independently managed from clusters and instances
in the very near future.