Make Identity Deletion Better (#179)
There is a race where stuff can be deleted in openstack, but still have a
record in the osi, so it'll attempt to delete it and fail, stuck in
error for all eternity.  We need to make openstack the source of truth,
so delete if we know about it, and allow 404s.

We need to be even better, what if something was provisioned but not
recorded, we should think about being able to lookup resources on
reconcile/delete rather than relying on the cache.