Add Identity References to Projects (#184)
Networks will very shortly be top level user created things, and as such
will need to be deleted on project deletion.  We need to inhibit project
deletion until we know all networks are cleaned up, so we add references
to the root node, the cloud identity, to pull this off.