Add References API to Network Resources
The file storage service is being migrated out of tree and thus will no
longer be allowed to directly manipulate network resources as that would
breach a security boundary.  This introduces a references API and client
that allows the file storage service to add and remove finalizers from
a network resource as required.

This follows the precedent set by the identity service and its project
references API thus maintaining consistency across all components.