Start Moving Off Stateful CRDs
When we create a resource, we then save state about the resource e.g.
it's ID, which we can then refer to later, either to prevent recreation
during reconcile, or to delete the resource.  Which is all well and good
until that state doesn't get saved, and you have resources floating
about that no one knows about.

The solution to this dilema is to not save state, and use OpenStack as
our source of truth.  To do this we need to be able to look for
resources to see whether they exist, and for that we need a predictable
but unique name.

This particular change implments the pattern for the networking layer.
For now, we will use the presence of an OpenstackNetwork resource to
trigger the execution of the legacy code path, just to keep the size
down.  Eventually we should be able to live migrate by reading the
OpenstackNetwork resources, and renaming all the resources before
finally deleting the OpenstackNetwork.