Rework Servers Logic
Much the same as before. However... it quickly became apparent that
servers refer to to networks and security groups and they may be in
either the old world or the new (and the same actually appplies to
security groups being part of networks).  This caused what can only be
described as spaghetti code that would have been next to impossible to
maintain.  So this deletes all the legacy stuff and goes all in on the
new/correct way of doing it.  The underlying OpenStack clients have been
made more type safe, thus limiting the scope for errors to a couple
places, rather than being sprinkled all over the provider.

This does however mean the next bit of work will be the migration task.