Openstack Resource Migration
The old way was keeping CRDs as records of resource IDs.  The new way is
to poll openstack using predictable names.  We need to migrate networks,
security groups and servers to this new methodology, which ostensibly
means delete the Openstack* CRDs.  Before we do this in a later release,
each controller needs to poll for it's managed Openstack CRD, do any
renames, and delete the resource.  An added complexity is that security
groups reference networks, and servers reference both, so inhibit
controller start up until the dependent resource controller have
performed their upgrades.