Fix Network Propagation
CAPO needs to have intrusive access to the OpenStack network and subnet
IDs, and now there is no way to get at this with the removal of the osn
resources.  For now, cache the results in the network status to make
them available over the API.  For now this is a private API, so safe,
but it's deprecated and will be removed eventually.  Much like the
identity API it'll probably need an admin route temporarily.  In the
future we should just create our own provider then we don't need to
reveal any internals and just pass around unikorn network IDs.