Servers V2 (#209)
Although the V1 API already does what we need it to, we can take this
opportunity to do some cleanups to the data structures, making it more
organized and simpler.  We do also need this API as the old one requires
identities, which will not be exposed to the higher order V2 services.