Fix Constant Reconcile in Compute (#214)
Rather naively we use reflect.DeepEqual to determine whether an update
is required, but in this case even encoding to JSON doesn't help.  What
we see is the required address pair has a CIDR and no MAC address,
however what's returned by the API actually has "" as the MAC e.g. it's
set irrespective of whether or not the underlying database entry is
populated.  Make the population conditional.