Simplify V2 APIs (#213)
For all the read, update delete APis we do a primary key lookup by ID,
then validate RBAC, so the organization/project ID parameters in the
paths are totally redundant and can be removed.  This just leaves
creation, which we handle now in the same way we do for other immutable
values e.g. network IDs, as body parameters.  This vastly simplifies the
APIs and code.  List APIs now use query parameters to scope responses.