Add tests for image upload
This mocks out the provider abstraction, so it can be used to check
that uploading a disk image works.

The mocks are generated in the step `make generate`. Calling that
updated the similarly generated file
pkg/providers/openstack/mock/interfaces.go too, so I've included that
here. In general the mock file changes are uninteresting.

Co-authored-by: Michael Bridgen <michael.bridgen@nscale.com>