Response with conflict when image not ready
It's now possible to "see" a disk image that is still pending (not
fully uploaded), so in principle there could be server requests that
mention an unusable disk image.

So: if you try to create or update a server using that image, it
should really respond with something other than 404 Not
Found. Conflict seems a reasonable alternative.

Co-authored-by: Michael Bridgen <michael.bridgen@nscale.com>