Make image upload/snapshot polling asynchronous
Previously, the public API was:

 * create an image, optionally providing a serverID of a server to
   snapshot (if not, the image is expected to be uploaded subsequently)
 * upload an image, either by posting the file, or by specifying a URL

However, the snapshot and upload from URL cases are similar to each
other, so it makes more sense to arrange the API this way:

 * create an image, supplying a serverID for snapshot, or a URL to
   fetch the image;
 * upload an image file

In the first case, this should also be asynchronous: it returns more
or less straight away, and completes the snapshot or fetch in the
background. This changes a few things about how the _provider_ API
works, too:

 - the CreateImageFromServer method returns after asking for the snapshot,
   rather than waiting for it to complete

 - images are no longer filtered for active status, because we want to
   know when e.g., an image is pending

 - FinalizeImage is renamed to PublishImage

 - There's now a hook for clearing the cache, since it'll need to be
   called by background processes when they complete.

Then the flow is the same for all the cases: create the image record
as pending, upload or snapshot the file, then publish the image record
(make it active).

Co-authored-by: Michael Bridgen <michael.bridgen@nscale.com>