Factor out reader prep entirely
This commit factors out a set of helpers for making readers, that's
enough to cover all the uses in the test cases. The end result is that
the test func doesn't need any nolint annotation, because it's simple
enough.

One side-effect: the test case for context cancellation no longer
passes, which it shouldn't because I removed the context-aware
reader. It was passing by coincidence because the "disk" context was
invalid anyway (making it fail as expected).