Give the upload context a chance to be cancelled
This commit adds in a check on every tar header, for the context
having been cancelled. This is as a kind of approximation to every
chunk; for legit uploads, most of the time will be spent in
`io.Copy(...)`, and this won't matter unless the connection is dropped
straight away (in which case it might make the difference between
cleaning up quickly and timing out).