Fill in image fetch and snapshot monitor tasks
This adapts the snapshot monitor and image upload code, as it was run
synchronously before (and in controllers in alansy/inst-72), into the
sagas.

The simplest thing is to run a goroutine in each case. This has
obvious flaws:

 - there's no rate limiting (could do this with a queue and workers);
 - there's no retries;
 - the goroutines either block graceful shutdown, or (with
   modifications) can be interrupted.

Any interruption would mean the tasks never complete. In the case of
snapshots, the image will still eventually be available, but it might
take a while for a server cache to pick it up. In the case of uploads,
that's the one shot gone.