Fix Server Delete Race
When the instance service tries to delete the node it adds the "wait for
children to vanish" finalizer, even though there are no children.  This
interestingly ended up in lockstep with the deprovisioning, instances
adding the finalizer, server controller not being able to update the
status because the resource version had changed, etc, etc.  Remove the
erroneous use of said finalizer.