Move provider implementations to internal/
It's easier to make changes if implementations are internal, and not
exported API. Certainly, consumers of this module should not be
calling into the providers directly.

To guard this, move the OpenStack and Kubernetes implementations into
`internal/`. This means a few imports change -- but only a few, most
packages did not import an implementation directly :tada:. Some
(generic) errors from OpenStack must move to the more providers/types/
package, since they are checked from other packages.