Extend Security Groups V2 (#228)
The main goal here is to enable ICMP, but while we are at it we can also
allow rules for any protocol, and also VRRP, which could be useful.
This reworks the API to be less tedious at the handler level, and also
removes a few contraints.  This allows you to "allow all from anywhere",
"allow all from a CIDR", "allow all TCP from anywhere" etc. etc.  This
also expands some error reporting to stop silly things like specifying a
port when you aren't using a layer 4 protocol.