Use a PVC in region server space to stage uploads
At present we're staging disk image uploads (either as posted, or from
a given URL) on disk. For a pod, that means it needs a writable volume
mounted. It's possible to mount ephemeral storage, which for some
$TMPDIR purposes would be fine; but the disk uploads are tens of
gigabytes, and that would blast through ephemeral storage capacity
quickly. So: use a persistent volume.

The storage class and size can be given in the chart values. Since
local deployments (on e.g., colima) may not have a storage class
supporting `ReadWriteMany`, there's a flag to use `ReadWriteOnce`
rather than `ReadWriteMany`.

(In a real cluster `ReadWriteOnce` would deadlock rolling upgrades --
a new pod can't start if the old pod hasn't released the volume; the
old pod can't terminate if the new pod hasn't started. But on a single
node, it's OK because pods on the same node can share a
`ReadWriteOnce` volume.)