Refactor: break down provisioners, provider interface, and region handler (#238)
* Break the Provider interface down

This means the clients in pkg/handlers can have more precise
requirements, which makes them easier to test and limits the blast
radius on changes.

* Move provider implementations to internal/

It's easier to make changes if implementations are internal, and not
exported API. Certainly, consumers of this module should not be
calling into the providers directly.

To guard this, move the OpenStack and Kubernetes implementations into
`internal/`. This means a few imports change -- but only a few, most
packages did not import an implementation directly :tada:. Some
(generic) errors from OpenStack must move to the more providers/types/
package, since they are checked from other packages.

* Factor out common provisioner shortcuts

The provisioner (managers) all use a similar "get client, get
identity, get provider" pattern. This factors some shortcuts out. As a
bonus I can remove a nolint on one of the provisioner methods.

* Use a more specific provider for server v2 handler

The types.Provider interface is used implicitly in many places, but
named mostly in the provisioners rather than the handlers. One place
it is named is in server v2 handlers, where it's used directly. For
the sake of precision and easier fake injection, it can be narrowed.

* Move conversion funcs out of region.go

This moves the flavor conversion funcs out of region, either into
`pkg/handler/conversion` so they can be used elsewhere, or into
`./conversion.go` for the sake of tidiness.

* Move image handler code out into its own package

This makes it easy to test on its own, and to add incrementally to the
handlers (e.g., for image upload) with less chance of conflicts.