Fix Empty Security Group Rule
The rules array is required, mainly as a convenience thing for the UX
side of things so you don't have to create and delete it, just push and
pop directly.  Unfortunately the read conversion spotted when the input
was empty and returned nil, rather than an empty array, which caused the
OpenAPI code to throw a wobbly.  Always allocate an array, even if
empty.