Handler and Client methods for new image paths
This adds Handler methods (expected by the generated code) for the new
API paths, and image.Client methods for them to call.

The Client methods are stubs, there to make this change
self-contained.

Mostly the handlers straight-forwardly adapt between HTTP payloads and
request values. The exception is the handler for uploading image data,
which has some defensive request size limitation (controlled by a
flag), and parses the request form before passing it on to the
client. I've left room here to handle other ways to upload the data,
e.g., as just the file bytes.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>