Extend Provider interface for image uploads
This commit adds the ImageWrite interface to the provider interfaces,
and extends the (provider) Image type.

The big change in the type is that images can now belong to an
organisation, and that they can have a status which might not be
"active". The status is because created images start in a queued
state, and callers will want to be able to tell pending images from
completed images.

The OpenStack implementation of the new interface gets a temporary
"not implemented", since it will have an implementation eventually;
the Kubernetes provider gets its standard ErrUnimplmented [sic].

This commit includes errors which the provider implementation will
want to use to abstract problems at the provider API.

Co-authored-by:	alansy-nscale <alan.sy@nscale.com>