Implement OpenStack client Image methods
The provider implementation adapts between the abstract Provider
interface, and the actual provider API (e.g., using OpenStack
clients). The precedent is to do this in two hops: first, implement
useful methods in a client wrapper; then, convert between the
abstracted values passed into the provider interface, and the client
methods.

This commit implements the ImageWrite methods for the OpenStack
client, and alters the ImageRead methods to suit the changed
expectations of the handler.

In particular:

 - images can now be tied to an organization, and if so, they are
   _not_ expected to be signed;

 - GetImage no longer depends on ListImages -- it checks the cache,
   and if not there, asks the upstream directly.

Several labels are now used on the write path as well as read, so I've
extracted them into consts.

The image cache moves to the provider, so that it's not discarded when
the image client is reconstructed.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>