Fix Security Group Delete Idempotency (#247)
What was seen in the wild was a security group stuck in deletion with it
unable to proceed due to the resource not being found.  In literally
every other handler this is ignored, and assumes that the resource has
been cleaned up.  In this instance a single ! missing.

I would add tests, but it's really complicated at this level to mock out
the OpenStack clients, so will add a backlog task to investigate and
action.  These will neet to apply to all top level reconcile and delete
functions for all resources, hence my reticence to tackle this now.