Implement Image client used by handler
The client used by the HTTP handler adapts between the requests and
the provider; the interfaces are fairly similar, but there's a bit of
validation and dispatch to do.

This commit fills in the client methods. The create image handler use
the saga pattern. There's really only one retractable operation
(creating an image record in the provider) for now, but since

 1. it can occur in the middle of the whole thing; and,
 2. at some point there will be allocations to allocate,

it's useful to be able to retract things.

This commit also adds tests, which use a mock to stand in for the
provider.

I've left room in `upload.go` for handling different forms the
uploaded file might come in. The initial form supported is a gzipped
tarball; we can extend this by swapping out the extraction step for
something else, e.g., just decompressing the input stream; and, we can
make decompression contingent on Content-Encoding, or the file name,
and MIME type (though more of those might need to be passed through
from the handler method).

Co-authored-by: alansy-nscale <alan.sy@nscale.com>