Make image handler more testable
This makes it possible to test the image (upload) handling in
isolation, by:

 - making a struct that just handles image requests (this involves
   moving the `setCacheable` method into *Options, so it's available
   to both handlers);
 - making the image client parameterised on how it obtains a region;
 - giving the image handler a region constructor to give to the
   client.

The effect is that (within the package at least) you can inject a mock
provider into the handler, and control what responses it will give.

The pkg/handler/image's provider type is the interface that's mocked;
it now needs to be exported, so it can be mentioned in tests outside
that package (i.e., of handlers).