Fix API Query Filtering
There was a bug in the label selector code whereby project labels were
not correctly matched by admin users, or at all for regular users with
the wrong project, which could lead to the project query being a no-op
and not actually working as advertised.  This picks up the new code.