Stable generated names for enum values (#242)
* Force oapi-codegen to use stable enum value names

Recently (6f64488), a change introduced a new enum value which was the
same as another enum name, and this cause oapi-codegen to give the
enum value names a prefix, to avoid a collision.

To avoid dumb breakages like this in the future, this commit adds a
flag to openapi/config.yaml which will force it to always prefix enum
value names. This avoid accidentally breaking things with an innocent
OpenAPI spec change; and makes it possible to break the schema out in
multiple files without those names reverting back to shorthands and
colliding again.

* Generate OpenAPI files when making on Mac

The recipes for bin/amd64... and bin/arm64 were different; the one for
arm64 did not have the generated OpenAPI files as a prerequisite. This
caused me confusion, building on my Mac, until I ran `make -d`.