Test good upload requests as well as bad
Adds tests for good requests, as well as some more ways you can send a
bad request. There's a couple more helpers factored out so they can be
used to create request and payloads.

One change to the code _under_ test is that I've made a failure to
find the right file in an upload a bad request, rather than a server
error. Say "No" to server errors!