Handle disk contents by themselves
This adds a small dispatch func so that people can upload just the
disk file, rather than a gzipped tarball.

Ultimately what we want to do with the data is the same, it's just
that it's decoded in different steps.

So: construct a callback for the data that will eventually be
uncovered, and pass that through the dispatch. This lets each decoding
step tidy up after itself, e.g., by closing readers after calling the
callback.

The tarball extraction doesn't bother copying its reader to a temp
file any more. It doesn't need to because it can stop when it sees the
right file name (before, it went on to validate that there were no
duplicate entries).

In general, I have just passed the decoded, possibly partial fetched
contents on as the request body to the provider API. This means the
fetch connection will stay open until the provider has received all
the data. This is the simplest way to do it, rather than the most
efficient.