Fix Monitor
It's entirely possible for a server to not have a backing node, at which
point the monitor will give up entirely and any servers in the queue
behind the failure will not get power status updates, for example never
turning on!  So Loosen the constraints a little by propagating not found
errors.  Best effort may be an option too, but let's see where this gets
us.