Filter out inactive images in list
The list images endpoint expects to be able to see the images
available to start instances with. It shouldn't see any that are not
yet usable.

Now there is the ability to create new images, which will start in a
pending or queued state; so we need to filter them from the list of
images.