API for create and upload custom images (#218)
* Move image handler to its own file

* Add API types/paths for image create

This adds components and paths to the OpenAPI spec for an image
creation API:

 * create an image record, supplying a URL to fetch the image
 * delete an image record.

The build will not build at this commit, because the handler lacks
methods implied by `server.spec.yaml`.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>

* Handler and Client methods for new image paths

This adds Handler methods (expected by the generated code) for the new
API paths, and image.Client methods for them to call.

The Client methods are stubs, there to make this change
self-contained.

Mostly the handlers straight-forwardly adapt between HTTP payloads and
request values.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>

* Extend Provider interface for image uploads

This commit adds the ImageWrite interface to the provider interfaces,
and extends the (provider) Image type.

The big change in the type is that images can now belong to an
organisation, and that they can have a status which might not be
"active". The status is because created images start in a queued
state, and callers will want to be able to tell pending images from
completed images.

The OpenStack implementation of the new interface gets a temporary
"not implemented", since it will have an implementation eventually;
the Kubernetes provider gets its standard ErrUnimplmented [sic].

This commit includes errors which the provider implementation will
want to use to abstract problems at the provider API.

Co-authored-by:	alansy-nscale <alan.sy@nscale.com>

* Implement Image client used by handler

The client used by the HTTP handler adapts between the requests and
the provider; the interfaces are fairly similar, but there's a bit of
validation and dispatch to do.

This commit fills in the client methods. The create image handler uses
the saga pattern. There's really only one retractable operation
(creating an image record in the provider) for now, but since

 1. it can occur in the middle of the whole thing; and,
 2. at some point there will be allocations to allocate,

it's useful to be able to retract things.

This commit also adds tests, which use a mock to stand in for the
provider.

I've left room in `upload.go` for handling different forms the
uploaded file might come in. The initial form supported is a gzipped
tarball; we can extend this by swapping out the extraction step for
something else, e.g., just decompressing the input stream; and, we can
make decompression contingent on Content-Encoding, or the file name,
and MIME type (though more of those might need to be passed through
from the handler method).

On reader_test.go:

We use Peek(), but I don't want to allocate the exact number needed
for that because it's likely to be too slow. I'd rather let the buffer
be the default size; but it needs to be bigger than we need, so we
don't just get an error. This test checks that assumption.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>

* Implement OpenStack client Image methods

The provider implementation adapts between the abstract Provider
interface, and the actual provider API (e.g., using OpenStack
clients). The precedent is to do this in two hops: first, implement
useful methods in a client wrapper; then, convert between the
abstracted values passed into the provider interface, and the client
methods.

This commit implements the ImageWrite methods for the OpenStack
client, and alters the ImageRead methods to suit the changed
expectations of the handler.

In particular:

 - images can now be tied to an organization, and if so, they are
   _not_ expected to be signed;

 - GetImage no longer depends on ListImages -- it checks the cache,
   and if not there, asks the upstream directly.

Several labels are now used on the write path as well as read, so I've
extracted them into consts.

The image cache moves to the provider, so that it's not discarded when
the image client is reconstructed.

Co-authored-by: alansy-nscale <alan.sy@nscale.com>

* Make image handler more testable

This makes it possible to test the image (upload) handling in
isolation, by:

 - making a struct that just handles image requests (this involves
   moving the `setCacheable` method into *Options, so it's available
   to both handlers);
 - making the image client parameterised on how it obtains a region;
 - giving the image handler a region constructor to give to the
   client.

The effect is that (within the package at least) you can inject a mock
provider into the handler, and control what responses it will give.

The pkg/handler/image's provider type is the interface that's mocked;
it now needs to be exported, so it can be mentioned in tests outside
that package (i.e., of handlers).

* Test image fetch handlers

Adds tests that the CreateImage endpoint will fetch the content of the
sourceURL given and upload that as the image data to a mock provider.

* Handle disk contents by themselves

This adds a small dispatch func so that people can upload just the
disk file, rather than a gzipped tarball.

Ultimately what we want to do with the data is the same, it's just
that it's decoded in different steps.

So: construct a callback for the data that will eventually be
uncovered, and pass that through the dispatch. This lets each decoding
step tidy up after itself, e.g., by closing readers after calling the
callback.

The tarball extraction doesn't bother copying its reader to a temp
file any more. It doesn't need to because it can stop when it sees the
right file name (before, it went on to validate that there were no
duplicate entries).

In general, I have just passed the decoded, possibly partial fetched
contents on as the request body to the provider API. This means the
fetch connection will stay open until the provider has received all
the data. This is the simplest way to do it, rather than the most
efficient.

* Filter out inactive images in list

The list images endpoint expects to be able to see the images
available to start instances with. It shouldn't see any that are not
yet usable.

Now there is the ability to create new images, which will start in a
pending or queued state; so we need to filter them from the list of
images.

---------

Co-authored-by: alansy-nscale <alan.sy@nscale.com>