Report the VLAN ID in the network status
This puts the VLAN ID in the OpenStack part of a Network status. That
part of the status is flagged for removal (I would guess in favour of
more generic status), but it is difficult to escape needing this
detail for getting anything done. In this case, it will supplying the
VLAN ID (and IP range) in the file storage resource, so it can be
attached to the network at the provider.