Populate file storage attachments from Network object
The Network object now has enough information in its status for the
storage client to fully populate its attachment fields. (These are
then used by the driver to attach storage to the right networks; but
the work here stops at getting the info into the custom resource.)

I have weakened the tests for validation of auth info and principal
slightly, because they are coupled to the order in which the
implementation attempts operations. Now that it generates the
attachments first, using the network handler, that will fail first --
so the error message will not be as it was expected.