Record the storage IP range when reconciling subnet
This feels like a better place to do it, since an IP range is an L3
thing.

It may not make much of a difference in practice; excepting that it
will now set it whether it not it created the subnet, whereas
previously it was only set on creation of the L2 network.