Enhance storage attachment with network configuration (#280)
* Report the VLAN ID in the network status

This puts the VLAN ID in the OpenStack part of a Network status. That
part of the status is flagged for removal (I would guess in favour of
more generic status), but it is difficult to escape needing this
detail for getting anything done. In this case, it will be supplying the
VLAN ID (and IP range) in the file storage resource, so it can be
attached to the network at the provider.

* Put the storage range in the network status too

Since the IP range is given in advance, the part of it dedicated to
storage can be calculated and reported too.

This makes it much easier to calculate the file storage attachments.

* Populate file storage attachments from Network object

The Network object now has enough information in its status for the
storage client to fully populate its attachment fields. (These are
then used by the driver to attach storage to the right networks; but
the work here stops at getting the info into the custom resource.)

I have weakened the tests for validation of auth info and principal
slightly, because they are coupled to the order in which the
implementation attempts operations. Now that it generates the
attachments first, using the network handler, that will fail first --
so the error message will not be as it was expected.

* Record the storage IP range when reconciling subnet

This feels like a better place to do it, since an IP range is an L3
thing.

It may not make much of a difference in practice; excepting that it
will now set it whether it not it created the subnet, whereas
previously it was only set on creation of the L2 network.

---------

Co-authored-by: Michael Bridgen <michael.bridgen@nscale.com>