Narrow the IP range used for attachments (#285)
* Narrow the IP range used for attachments

The provider carves out a large range of IPs, and we don't need them all.
So: just use the first four.

This updates the test to give the network a large range, and check the
attachment gets a narrow range, as desired.

* Avoid indirecting through nil storageRange

Previously, the storage range value was used as-is, so it didn't
matter if it was `nil`; now, we indirect through it, so it better not
be nil.